support both modes of plugin activation
